+++ db/migrate/20190220235611_create_table_orderinv.rb
@@ -2,7 +2,7 @@ class CreateTableOrderinv < ActiveRecord::Migration[5.2]
   def change
   	create_table :order_invs do |t|
 	  	t.string :invoice, null: false
-	  	t.integer :nominal, null: false
+	  	t.bigint :nominal, null: false
 	  	t.references :order, foreign_key: true, null:false
 	  	t.datetime :date_paid
