+++ db/migrate/20190228125359_create_paid_pay_inv.rb
@@ -4,7 +4,7 @@ class CreatePaidPayInv < ActiveRecord::Migration[5.2]
     	t.string :invoice, null: false
     	t.integer :transaction_type, null: false, default: 1
     	t.string :transaction_invoice, null: false
-    	t.integer :nominal, null: false, default: 0
+    	t.bigint :nominal, null: false, default: 0
     	t.timestamp :date_created, default: "CURRENT_TIMESTAMP"
         t.references :user, foreign_key: true, null: false
