+++ db/migrate/20190305084609_create_transaction_item.rb
@@ -3,9 +3,9 @@ class CreateTransactionItem < ActiveRecord::Migration[5.2]
     create_table :transaction_items do |t|
     	t.references :item, foreign_key: true, null: false
     	t.references :transaction, foreign_key: true, null: false
-    	t.integer :price, null: false
-    	t.integer :discount, default: 0
-    	t.integer :quantity, null: false
+    	t.bigint :price, null: false
+    	t.bigint :discount, default: 0
+    	t.bigint :quantity, null: false
     	t.datetime :date_created
         t.integer :retur
         t.integer :replace
